+++ safeincave/Materials/Constitutive/ViscoplasticDesai.py
@@ -5,7 +5,7 @@
 from __future__ import annotations
 import torch as to
 import numpy as np
-from ..Utils import MPa
+from ...Utils import MPa
 from .NonElasticElement import NonElasticElement
